# src/apiDocs/openapi.yaml
openapi: 3.0.3
info:
  title: ELanguageCenter IELTS B2B API
  description: This is the API definition for ELanguageCenter IELTS B2B(Business to Business) API
  termsOfService: http://swagger.io/terms/
  contact:
    email: elanguagecenter23@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:8000
paths:
...
  /ielts/test/{testId}/speaking:
    post:
      summary: Create new speaking test
      description: Create a new IELTS speaking test for a user
      operationId: createNewSpeakingTest
      tags:
        - Speaking
      parameters:
        - in: path
          name: testId
          required: true
          schema:
            type: string
      requestBody:
        description: New speaking test request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/speakingTestRequest'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/speakingTestResponse'
        '400':
          description: Invalid request in the payload
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ELCITELTSErrorResponse'
        '401':
          description: User is unauthorized
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ELCITELTSErrorResponse'
        '500':
          description: Internal server error
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Headers:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ELCITELTSErrorResponse'
...
components:
  schemas:
...
    speakingTestRequest:
      type: object
      properties:
        name:
          type: string
          example: SpeakingTest1
    speakingTestResponse:
      type: object
      properties:
        id:
          type: string
          example: 19c4bcd9-9bc3-4d65-a853-0a6081baad0a
        name:
          type: string
          example: SpeakingTest1
          
    # Speaking Test Stage data types
...
    ELCITELTSErrorResponse:
      type: object
      properties:
        code:
          type: number
          example: 400
        message:
          type: string
          example: Invalid test name
  securitySchemes:
    bearerAuth:  # Moved the section to align with components
      type: http
      scheme: bearer